adding changes

+++ InvenShopfy.Core/Requests/Products/Product/CreateProductRequest.cs
@@ -25,8 +25,7 @@ public class CreateProductRequest : Request
     public string ProductImage { get; set; } = null!;
     
     [Required(ErrorMessage = "Invalid Quantity")]
-    
-    public double Quantity { get; set; }
+    public long Quantity { get; set; }
     
     [Required(ErrorMessage = "Invalid Unit Id")]
     public long UnitId  { get; set; }
