+++ InvenShopfy.Core/Requests/Products/Product/UpdateProductRequest.cs
@@ -25,7 +25,7 @@ public class UpdateProductRequest : Request
     
     
     [Required(ErrorMessage = "Invalid Quantity")]
-    public double Quantity { get; set; }
+    public long Quantity { get; set; }
     
     // [Required(ErrorMessage = "Invalid Image")] // NEED TO COME BACK AS I WILL BE ADDING THIS TO CLOUDNIARY
     public string ProductImage { get; set; } = null!;
